requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bcrypt==5.0.0
click==8.3.1
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.124.4
h11==0.16.0
idna==3.11
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.15.5
python-dotenv==1.2.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
